requirements.txt
altgraph==0.17.4
black==24.1.0
click==8.1.7
colorama==0.4.6
convertdate==2.4.0
customtkinter==5.2.2
darkdetect==0.8.0
fpdf==1.7.2
hijri-converter==2.2.4
holidays==0.16
korean-lunar-calendar==0.2.1
mypy-extensions==1.0.0
packaging==23.2
pathspec==0.12.1
pefile==2023.2.7
pillow==10.2.0
platformdirs==4.1.0
pyinstaller==6.3.0
pyinstaller-hooks-contrib==2023.12
PyMeeus==0.5.11
python-dateutil==2.8.2
pywin32-ctypes==0.2.2
six==1.16.0
sqlitedict==2.0.0
XlsxWriter==3.0.3
